+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.3.9 [unreleased]
 
-- Fixed issue with non-lowercase tables and partial syncs
+- Better support for schemas
 - Added `--schema` option to sync with schema
+- Fixed issue with non-lowercase tables and partial syncs
 
 # 0.3.8